requirements.txt
streamlit==1.28.1
openai==1.12.0
httpx>=0.24.0
python-dotenv==1.0.0
pandas==2.1.3
streamlit-chat==0.1.1
